SATE/requirements.txt
et-xmlfile==1.1.0
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scipy==1.7.3
six==1.16.0
xlrd==2.0.1